refactor out bulk of logic into base hook so it can be re-used
