+++ src/utils/types/enums/WindowClosingOptions.ts
@@ -1,0 +1,6 @@
+enum ClosingOption {
+    AllChildren,
+    AnyChild,
+}
+
+export default ClosingOption;
